% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_descriptor.R
\name{score_descriptor}
\alias{score_descriptor}
\title{#' Generate Descriptive Labels for Standardized Test Scores}
\usage{
score_descriptor(score, metric = "index", system = "aan", keep_na = TRUE)
}
\arguments{
\item{score}{Numeric value (or vector) containing the standardised scores to
be converted.}

\item{metric}{Character string specifying the metric of the input scores.
("z", "t", "index", "scaled", "sten", "stanine", "sat"). Default is "index".}

\item{system}{Character string specifying the descriptor system to use.
Options include "aan", "wisc", "wais", "groth.marnat", "nepsy", "q.simple",
and "heaton". Default is "aan".}

\item{keep_na}{NA values are kept as NA without throwing a warning message.}
}
\value{
A character vector containing the qualitative descriptors (e.g.,
"Low", "Average", "High") corresponding to the input test scores based on
the selected descriptor system.
}
\description{
This function assigns qualitative descriptors to standardized test scores
based on predefined systems. Descriptors are derived according to established
thresholds in various neuropsychological assessment systems (see details for
more information).
}
\details{
The function uses predefined threshold values and descriptors from
various neuropsychological assessment systems to convert test scores into
qualitative categories. By default, it uses the system from the American
Academy of Clinical Neuropsychology consensus paper (Guilmette et al.,
2020).

Available descriptor systems include Guilmette et al. (2020), Wechsler
WISC-V (2014), Wechsler WAIS-IV (2008), NEPSY II (2007), Q Simple (2017),
Groth-Marnat (2009), and Heaton (1991). The preferred and default system for
this function is the American Academy of Clinical Neuropsychology consensus
(Guilmette et al., 2020). For details on other descriptor systems, see
Schoenberg and Rum (2017).
}
\examples{
# Convert index scores to descriptors using the AAN descriptor system
score_descriptor(score = c(85, 105, 120), metric = "index", system = "aan")

# Convert index scores to descriptors using the NEPSY descriptor system
score_descriptor(score = c(75, 95, 110), metric = "index", system = "nepsy")

}
\references{
\itemize{
\item Schoenberg, M.R., & Rum, R.S. (2017). Towards reporting standards for
neuropsychological study results: A proposal to minimize communication
errors with standardized qualitative descriptors for normalized test scores.
\emph{Clinical Neurology and Neurosurgery, 162}, 72-79.
\item Guilmette, T.J., Sweet, J.J., Hebben, N., Koltai, D., Mahone, E.M.,
Spiegler, B.J., Stucky, K., & Westerveld, M. (2020).
American Academy of Clinical Neuropsychology consensus conference
statement on uniform labeling of performance test scores.
\emph{The Clinical Neuropsychologist, 34}(3), 437-453.
doi:10.1080/13854046.2020.1722244.
}
}
\seealso{
\itemize{
\item \code{\link[=convert_z]{convert_z()}}: for converting z scores to other standardised metrics.
\item \code{\link[=convert_standard]{convert_standard()}}: For assessing a dissociation between two test scores for a single case.
}
}
